+++ src/js/pages/ManageMyCandidates/ManageMyCandidatesLanding.jsx
@@ -1,13 +1,14 @@
-import React, { Suspense, useCallback, useEffect, useMemo, useState, useRef } from 'react';
-import { useHistory, useParams } from 'react-router-dom';
-import styled from 'styled-components';
-import { Helmet } from 'react-helmet-async';
 import { Edit as EditIcon, KeyboardArrowDown as ArrowDownIcon } from '@mui/icons-material';
 import PropTypes from 'prop-types';
-import PoliticianStore from '../../common/stores/PoliticianStore';
-import VoterStore from '../../stores/VoterStore';
-import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
+import React, { Suspense, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { Helmet } from 'react-helmet-async';
+import { useHistory, useParams } from 'react-router-dom';
+import styled from 'styled-components';
 import DesignTokenColors from '../../common/components/Style/DesignTokenColors';
+import PoliticianStore from '../../common/stores/PoliticianStore';
+import { PageContentContainer } from '../../components/Style/pageLayoutStyles';
+import VoterStore from '../../stores/VoterStore';
+import ImportInviteIcon from '../../components/More/ImportInviteIcon';
 
 
 const PoliticiansManagedController = React.lazy(() => import('../../components/PoliticiansManaged/PoliticiansManagedController'));
@@ -257,26 +258,6 @@ const TrackingIcon = ({ size = 22, title = 'Tracking', ...props }) => (
       fillRule="evenodd"
       clipRule="evenodd"
       d="M16.2727 2.02341C16.3823 2.05363 16.482 2.1124 16.5615 2.19371C16.641 2.27503 16.6975 2.37598 16.7252 2.48626L17.2742 4.68483L19.4728 5.23383C19.5831 5.26159 19.6841 5.31813 19.7655 5.39768C19.8468 5.47722 19.9056 5.57693 19.9358 5.68662C19.966 5.7963 19.9665 5.91203 19.9374 6.022C19.9082 6.13197 19.8504 6.23223 19.7698 6.31255L17.1984 8.88398C17.1199 8.96218 17.0226 9.01878 16.9158 9.04829C16.809 9.0778 16.6964 9.07922 16.589 9.05241L14.5524 8.54455L11.5271 11.5685C11.3764 11.7193 11.172 11.8039 10.9588 11.8039C10.7457 11.8039 10.5412 11.7193 10.3905 11.5685C10.2398 11.4178 10.1551 11.2134 10.1551 11.0003C10.1551 10.7871 10.2398 10.5827 10.3905 10.432L13.4145 7.40669L12.9054 5.37012C12.8783 5.26244 12.8796 5.1496 12.9091 5.04257C12.9387 4.93555 12.9954 4.83799 13.0738 4.75941L15.6452 2.18798C15.7257 2.10761 15.8261 2.05003 15.9361 2.02112C16.0461 1.9922 16.1618 1.99299 16.2714 2.02341M10.4471 3.64341C10.6599 3.62908 10.8583 3.53082 10.9986 3.37024C11.1389 3.20965 11.2097 2.9999 11.1954 2.78712C11.1811 2.57434 11.0828 2.37597 10.9222 2.23563C10.7616 2.0953 10.5519 2.02451 10.3391 2.03883C8.62767 2.15797 6.98641 2.76586 5.61028 3.79027C4.23415 4.81469 3.18099 6.21259 2.57591 7.8179C1.97082 9.42322 1.83924 11.1685 2.1968 12.8464C2.55436 14.5243 3.38603 16.0643 4.59301 17.2834C5.79999 18.5026 7.33156 19.3497 9.00578 19.7241C10.68 20.0984 12.4265 19.9844 14.0378 19.3955C15.6491 18.8065 17.0575 17.7675 18.0957 16.4017C19.1339 15.0359 19.7582 13.4008 19.8945 11.6907C19.9116 11.4783 19.8435 11.2677 19.7054 11.1055C19.5672 10.9432 19.3702 10.8425 19.1578 10.8254C18.9454 10.8084 18.7349 10.8764 18.5726 11.0146C18.4103 11.1527 18.3096 11.3497 18.2925 11.5621C18.1811 12.966 17.669 14.3083 16.817 15.4297C15.9651 16.551 14.8092 17.4042 13.4866 17.888C12.1641 18.3717 10.7304 18.4656 9.35607 18.1585C7.9817 17.8514 6.72437 17.1563 5.73346 16.1556C4.74256 15.155 4.05973 13.8909 3.76611 12.5136C3.47249 11.1363 3.58042 9.70363 4.07708 8.38585C4.57373 7.06808 5.43823 5.92057 6.56787 5.07967C7.69751 4.23878 9.0448 3.73983 10.4497 3.64212M10.3275 7.20355C10.366 7.30181 10.3848 7.4067 10.3828 7.51222C10.3807 7.61774 10.3579 7.72183 10.3157 7.81854C10.2734 7.91525 10.2125 8.00268 10.1364 8.07585C10.0604 8.14902 9.97066 8.20649 9.87238 8.24498C9.41381 8.42437 9.00719 8.71524 8.6893 9.09129C8.37141 9.46734 8.15228 9.9167 8.05172 10.3987C7.95116 10.8808 7.97235 11.3803 8.11337 11.852C8.25439 12.3238 8.51079 12.753 8.85938 13.1008C9.20798 13.4486 9.63777 13.704 10.1099 13.8439C10.582 13.9838 11.0815 14.0038 11.5633 13.9021C12.0451 13.8004 12.494 13.5802 12.8693 13.2615C13.2446 12.9427 13.5345 12.5354 13.7128 12.0764C13.7935 11.8829 13.9466 11.7286 14.1395 11.6464C14.3323 11.5642 14.5497 11.5607 14.7452 11.6365C14.9406 11.7123 15.0987 11.8615 15.1857 12.0523C15.2726 12.2431 15.2816 12.4603 15.2107 12.6575C14.936 13.3662 14.489 13.9953 13.91 14.4877C13.3311 14.9801 12.6385 15.3204 11.8949 15.4777C11.1513 15.635 10.3802 15.6044 9.65142 15.3887C8.92263 15.173 8.25914 14.7789 7.72101 14.2422C7.18287 13.7054 6.78708 13.043 6.56944 12.3148C6.35181 11.5865 6.31921 10.8155 6.4746 10.0715C6.62998 9.32754 6.96845 8.63404 7.45935 8.0538C7.95026 7.47355 8.57812 7.02489 9.2861 6.74841C9.38436 6.7099 9.48925 6.69112 9.59477 6.69316C9.70029 6.69519 9.80438 6.71799 9.90109 6.76026C9.9978 6.80252 10.0852 6.86342 10.1584 6.93948C10.2316 7.01555 10.289 7.10527 10.3275 7.20355Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-// Import & invite icon
-export const ImportInviteIcon = ({ size = 22, title = 'Import & invite', ...props }) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 22 22"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    role="img"
-    aria-hidden={title ? undefined : true}
-    {...props}
-  >
-    {title ? <title>{title}</title> : null}
-    <path
-      d="M16.8333 15V16.6667H6.83333V15C6.83333 15 6.83333 11.6667 11.8333 11.6667C16.8333 11.6667 16.8333 15 16.8333 15ZM14.3333 7.5C14.3333 7.00555 14.1867 6.5222 13.912 6.11108C13.6373 5.69995 13.2469 5.37952 12.79 5.1903C12.3332 5.00108 11.8306 4.95157 11.3456 5.04804C10.8607 5.1445 10.4152 5.3826 10.0656 5.73223C9.71593 6.08187 9.47783 6.52732 9.38137 7.01228C9.28491 7.49723 9.33441 7.99989 9.52363 8.45671C9.71285 8.91352 10.0333 9.30397 10.4444 9.57867C10.8555 9.85338 11.3389 10 11.8333 10C12.4964 10 13.1323 9.73661 13.6011 9.26777C14.0699 8.79893 14.3333 8.16304 14.3333 7.5ZM17 11.7167C17.4555 12.1369 17.8228 12.6437 18.0803 13.2074C18.3379 13.7712 18.4805 14.3805 18.5 15V16.6667H21V15C21 15 21 12.125 17 11.7167ZM16 5C15.7482 5.00002 15.498 5.03938 15.2583 5.11667C15.7459 5.81581 16.0073 6.64765 16.0073 7.5C16.0073 8.35235 15.7459 9.18419 15.2583 9.88333C15.498 9.96062 15.7482 9.99999 16 10C16.663 10 17.2989 9.73661 17.7678 9.26777C18.2366 8.79893 18.5 8.16304 18.5 7.5C18.5 6.83696 18.2366 6.20107 17.7678 5.73223C17.2989 5.26339 16.663 5 16 5ZM7.66667 9.16667H5.16667V6.66667H3.5V9.16667H1V10.8333H3.5V13.3333H5.16667V10.8333H7.66667V9.16667Z"
       fill="currentColor"
     />
   </svg>
